requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Classy==0.6.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pathlib==1.0.1
pyaml==20.4.0
PyYAML==5.4.1
termcolor==1.1.0
Werkzeug==1.0.1
